python-agent/requirements.txt
python-dotenv
openai
prometheus-client
opensearch-py
schedule
fastapi
uvicorn

# LangChain & Vector DB
langchain
chromadb